shopCart logic changed

+++ front/src/pages/Store/Shop/index.tsx
@@ -11,7 +11,7 @@ import api from "src/services";
 import { SHOP_ADD_ITEM_URL, SHOP_DELETE_ITEM_URL } from "../api.data";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { ICartItem, IShopData } from "../api.types";
-import { CourseAvailability } from "./types";
+import { ShopCourseStatus } from "./types";
 import { error } from "console";
 const BgList = [ShopItemBG1, ShopItemBG2, ShopItemBG3];
 
@@ -25,60 +25,24 @@ const Shop: React.FC = ({}) => {
     userData,
     isUserLoading,
   } = useContext(StoreContext);
+
   const [itemsAvailabilityList, setItemsAvailabilityList] = useState<
-    Array<CourseAvailability>
+    Array<ShopCourseStatus>
   >([]);
 
-  const onQuantityChange = (itemId: string, index: number) => {
-    return () => {
-      api.post(`${SHOP_ADD_ITEM_URL}/${itemId}`)({});
-      updateStatus(index);
-    };
+  const onQuantityChange = (action: "add" | "delete") => {
+    console.log("PASSED FROM CHANGE");
+    if (action === "add")
+      return (itemId: string) => {
+        console.log("PASSED FROM ADD");
+        api.post(`${SHOP_ADD_ITEM_URL}/${itemId}`)({});
+      };
+    else
+      return (itemId: string) => {
+        console.log("PASSED FROM DELETE");
+        api.delete(`${SHOP_ADD_ITEM_URL}/${itemId}`)({});
+      };
   };
-
-
-  const updateStatus = (itemIndex: number) => {
-    const newStatus = itemsAvailabilityList.map((item, index) =>
-      index === itemIndex
-        ? item === "purchased"
-          ? "purchased"
-          : item === "selected"
-          ? "available"
-          : "selected"
-        : item
-    );
-    setItemsAvailabilityList(newStatus);
-  };
-  const addToCart = (itemId: string, index: number) => {
-    api.post(`${SHOP_ADD_ITEM_URL}/${itemId}`)({});
-    updateStatus(index);
-  };
-  const deleteFromCart = (itemId: string, index: number): void => {
-    api.delete(`${SHOP_ADD_ITEM_URL}/${itemId}`)({});
-    updateStatus(index);
-  };
-  const findItemStatus = (shopData: IShopData, cartData: ICartItem[]) => {
-    const cartCoursesId = cartData.map((item) => item.product.id);
-    const courseAvailability = shopData.courses.map((course) => {
-      if (course.hasAccess) {
-        return "purchased";
-      } else if (cartCoursesId.includes(course.id)) {
-        return "selected";
-      } else {
-        return "available";
-      }
-    });
-    return courseAvailability;
-  };
-  useEffect(() => {
-    if (shopData && cartData) {
-      const shopItemsAvailabilityStatusList = findItemStatus(
-        shopData,
-        cartData
-      );
-      setItemsAvailabilityList(shopItemsAvailabilityStatusList);
-    }
-  }, [shopData, cartData]);
 
   return (
     <main className="max-w-[100vw] w-[100vw] min-w-[100vw]">
@@ -100,17 +64,9 @@ const Shop: React.FC = ({}) => {
           ) : isShopSuccess && shopData ? (
             shopData.courses.map((item, index) => (
               <ShopCourseCard
-                level={item.detail.level}
                 id={item.id}
                 key={item.id}
-                title={item.detail.title}
-                index={index}
-                price={item.detail.price}
-                itemStatus={itemsAvailabilityList[index]}
-                backgroundTuhmbURL={BgList[item.detail.level + 1]}
-                // addToCart={addToCart}
-                onQuantityChange={() => onQuantityChange}
-                // deleteFromCart={deleteFromCart}
+                onQuantityChange={onQuantityChange}
               />
             ))
           ) : (
